feat(cli): Make v2 of the resolver the default implementation
- Separates concerns: Building a dependency tree, rule based validation, and topological sorting are now separate
- Mutating CLI operations (like install or publish) are not required to happen in the recursive function parsing the dependency tree. Instead, they operate in-order on an ordered Vector of dependencies retrieved by DependencyGraph
- Separation of concerns allows for better unit testability of the code, relying less on finicky e2e tests
